clase4/pyproject.toml
[tool.poetry]
name = "clase4"
version = "0.1.0"
description = ""
authors = ["RArielAlonso <rariel.alonso@gmail.com>"]
readme = "README.md"
packages = [{include="example_pytest"} ,
{include="docstrings"},
{include= "pda" },
{include= "config_example" } ]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pdbpp = "^0.10.3"
pytest = "^7.4.2"
jupyter = "^1.0.0"
numpy = "^1.26.0"
pandas = "^2.1.1"
pydocstyle = "^6.3.0"
mypy = "0.900"
sphinx = "^7.2.6"
python-dotenv = "0.9"
pyyaml = "^6.0.1"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
